# docs/api/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/tags/{id}":{"get":{"tags":["tag-controller"],"operationId":"getTagById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponse"}}}}}},"put":{"tags":["tag-controller"],"operationId":"updateTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponse"}}}}}},"delete":{"tags":["tag-controller"],"operationId":"deleteTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductAdminUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/product-types/{id}":{"get":{"tags":["product-type-controller"],"operationId":"getProductTypeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductTypeResponse"}}}}}},"put":{"tags":["product-type-controller"],"operationId":"updateProductType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductTypeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductTypeResponse"}}}}}},"delete":{"tags":["product-type-controller"],"operationId":"deleteProductType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/clients/{id}":{"get":{"tags":["client-controller"],"operationId":"getClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}}},"put":{"tags":["client-controller"],"operationId":"updateClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}}},"delete":{"tags":["client-controller"],"operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/clients/profile":{"get":{"tags":["client-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientProfileResponse"}}}}}},"put":{"tags":["client-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientProfileResponse"}}}}}}},"/api/clients/password":{"put":{"tags":["client-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/brands/{id}":{"get":{"tags":["brand-controller"],"operationId":"getBrandById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandResponse"}}}}}},"put":{"tags":["brand-controller"],"operationId":"updateBrandAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandAdminUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandResponse"}}}}}},"delete":{"tags":["brand-controller"],"operationId":"deleteBrand","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/brands/profile":{"get":{"tags":["brand-controller"],"operationId":"getProfile_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandProfileResponse"}}}}}},"put":{"tags":["brand-controller"],"operationId":"updateProfile_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandProfileResponse"}}}}}}},"/api/brands/password":{"put":{"tags":["brand-controller"],"operationId":"changePassword_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/brands/my-products/{id}":{"get":{"tags":["brand-controller"],"operationId":"getMyProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["brand-controller"],"operationId":"updateMyProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["brand-controller"],"operationId":"deleteMyProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/admin/users/{id}/status":{"put":{"tags":["admin-controller"],"operationId":"setUserStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"active","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"}}}},"/api/admin/brands/{id}/verify":{"put":{"tags":["admin-controller"],"operationId":"verifyBrand","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"verified","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/tags":{"get":{"tags":["tag-controller"],"operationId":"getAllTags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagResponse"}}}}}}},"post":{"tags":["tag-controller"],"operationId":"createTag","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagResponse"}}}}}}},"/api/products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/products/search/reindex":{"post":{"tags":["product-search-controller"],"operationId":"reindexAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/product-types":{"get":{"tags":["product-type-controller"],"operationId":"getAllProductTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductTypeResponse"}}}}}}},"post":{"tags":["product-type-controller"],"operationId":"createProductType","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductTypeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductTypeResponse"}}}}}}},"/api/clients":{"get":{"tags":["client-controller"],"operationId":"getAllClients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientResponse"}}}}}}},"post":{"tags":["client-controller"],"operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}}}},"/api/brands":{"get":{"tags":["brand-controller"],"operationId":"getAllBrands","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BrandResponse"}}}}}}},"post":{"tags":["brand-controller"],"operationId":"createBrand","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandResponse"}}}}}}},"/api/brands/my-products":{"get":{"tags":["brand-controller"],"operationId":"getMyProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"post":{"tags":["brand-controller"],"operationId":"createMyProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/products/search":{"get":{"tags":["product-search-controller"],"operationId":"searchProducts","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductSearchResponse"}}}}}}},"/api/products/search/elastic-search":{"get":{"tags":["product-search-controller"],"operationId":"searchProducts_1","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"productTypeName","in":"query","required":false,"schema":{"type":"string"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}},{"name":"generalFit","in":"query","required":false,"schema":{"type":"string"}},{"name":"gender","in":"query","required":false,"schema":{"type":"string"}},{"name":"color","in":"query","required":false,"schema":{"type":"string"}},{"name":"available","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"minPrice","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"maxPrice","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"tags","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":12}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductElasticSearchResponse"}}}}}}},"/api/brands/public":{"get":{"tags":["brand-controller"],"operationId":"getAllPublicBrands","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BrandPublicResponse"}}}}}}}},"/api/brands/public/{id}":{"get":{"tags":["brand-controller"],"operationId":"getPublicBrandById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BrandPublicResponse"}}}}}}}},"components":{"schemas":{"UserUpdateRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"role":{"type":"string","enum":["CLIENT","BRAND","ADMIN"]},"isActive":{"type":"boolean"}},"required":["email","isActive","role"]},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"role":{"type":"string","enum":["CLIENT","BRAND","ADMIN"]},"createdAt":{"type":"string","format":"date-time"},"isActive":{"type":"boolean"}}},"TagRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":0},"type":{"type":"string","enum":["STYLE","OCCASION","FIT"]}},"required":["name","type"]},"TagResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string","enum":["STYLE","OCCASION","FIT"]}}},"ProductAdminUpdateRequest":{"type":"object","properties":{"brandId":{"type":"integer","format":"int64"},"productTypeId":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string"},"price":{"type":"number","minimum":0.0},"generalFit":{"type":"string","enum":["COMPRESSION","SKINNY","SLIM","REGULAR","RELAXED","LOOSE","OVERSIZED","OTHER"]},"gender":{"type":"string","enum":["MALE","FEMALE","UNISEX"]},"color":{"type":"string","enum":["WHITE","BLACK","GREY","BROWN","BEIGE","CREAM","GREEN","BLUE","NAVY","TURQUOISE","PURPLE","RED","MAROON","ORANGE","PINK","YELLOW","GOLD","SILVER","MULTICOLOR","OTHER"]},"available":{"type":"boolean"},"linkProduct":{"type":"string","maxLength":500,"minLength":0},"imageUrls":{"type":"array","items":{"type":"string"}},"tagIds":{"type":"array","items":{"type":"integer","format":"int64"},"uniqueItems":true}},"required":["brandId","name","price","productTypeId"]},"ProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"brandId":{"type":"integer","format":"int64"},"productType":{"$ref":"#/components/schemas/ProductTypeResponse"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"generalFit":{"type":"string","enum":["COMPRESSION","SKINNY","SLIM","REGULAR","RELAXED","LOOSE","OVERSIZED","OTHER"]},"gender":{"type":"string","enum":["MALE","FEMALE","UNISEX"]},"color":{"type":"string","enum":["WHITE","BLACK","GREY","BROWN","BEIGE","CREAM","GREEN","BLUE","NAVY","TURQUOISE","PURPLE","RED","MAROON","ORANGE","PINK","YELLOW","GOLD","SILVER","MULTICOLOR","OTHER"]},"available":{"type":"boolean"},"rating":{"type":"number","format":"double"},"linkProduct":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"imageUrls":{"type":"array","items":{"type":"string"}},"tags":{"type":"array","items":{"$ref":"#/components/schemas/TagResponse"}}}},"ProductTypeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"category":{"type":"string","enum":["TOPS","BOTTOMS","FOOTWEAR","OUTERWEAR","ACCESSORIES","FULL_BODY"]}}},"ProductTypeRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"category":{"type":"string"}},"required":["category","name"]},"ClientUpdateRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0}},"required":["name"]},"ClientResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ClientProfileUpdateRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0}},"required":["name"]},"ClientProfileResponse":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}}},"PasswordChangeRequestDTO":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}}},"BrandAdminUpdateRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"pictureURL":{"type":"string","maxLength":500,"minLength":0},"linkOfficial":{"type":"string","maxLength":255,"minLength":0},"isVerified":{"type":"boolean"}},"required":["isVerified","name"]},"BrandResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"pictureURL":{"type":"string"},"linkOfficial":{"type":"string"},"followers":{"type":"integer","format":"int32"},"rating":{"type":"number","format":"double"},"isVerified":{"type":"boolean"}}},"BrandProfileUpdateRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"pictureURL":{"type":"string","maxLength":500,"minLength":0},"linkOfficial":{"type":"string","maxLength":255,"minLength":0}},"required":["name"]},"BrandProfileResponse":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"pictureURL":{"type":"string"},"linkOfficial":{"type":"string"},"followers":{"type":"integer","format":"int32"},"rating":{"type":"number","format":"double"},"isVerified":{"type":"boolean"}}},"ProductUpdateRequest":{"type":"object","properties":{"productTypeId":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string"},"price":{"type":"number","minimum":0.0},"generalFit":{"type":"string","enum":["COMPRESSION","SKINNY","SLIM","REGULAR","RELAXED","LOOSE","OVERSIZED","OTHER"]},"gender":{"type":"string","enum":["MALE","FEMALE","UNISEX"]},"color":{"type":"string","enum":["WHITE","BLACK","GREY","BROWN","BEIGE","CREAM","GREEN","BLUE","NAVY","TURQUOISE","PURPLE","RED","MAROON","ORANGE","PINK","YELLOW","GOLD","SILVER","MULTICOLOR","OTHER"]},"available":{"type":"boolean"},"linkProduct":{"type":"string","maxLength":500,"minLength":0},"imageUrls":{"type":"array","items":{"type":"string"}},"tagIds":{"type":"array","items":{"type":"integer","format":"int64"},"uniqueItems":true}},"required":["name","price","productTypeId"]},"UserCreateRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6},"role":{"type":"string","enum":["CLIENT","BRAND","ADMIN"]}},"required":["email","password","role"]},"ProductCreateRequest":{"type":"object","properties":{"brandId":{"type":"integer","format":"int64"},"productTypeId":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":150,"minLength":0},"description":{"type":"string"},"price":{"type":"number","minimum":0.0},"generalFit":{"type":"string","enum":["COMPRESSION","SKINNY","SLIM","REGULAR","RELAXED","LOOSE","OVERSIZED","OTHER"]},"gender":{"type":"string","enum":["MALE","FEMALE","UNISEX"]},"color":{"type":"string","enum":["WHITE","BLACK","GREY","BROWN","BEIGE","CREAM","GREEN","BLUE","NAVY","TURQUOISE","PURPLE","RED","MAROON","ORANGE","PINK","YELLOW","GOLD","SILVER","MULTICOLOR","OTHER"]},"available":{"type":"boolean"},"linkProduct":{"type":"string","maxLength":500,"minLength":0},"imageUrls":{"type":"array","items":{"type":"string"}},"tagIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"required":["name","price","productTypeId"]},"ClientCreateRequest":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":100,"minLength":0}},"required":["name","userId"]},"BrandCreateRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"pictureURL":{"type":"string","maxLength":500,"minLength":0},"linkOfficial":{"type":"string","maxLength":255,"minLength":0}},"required":["name"]},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6},"role":{"type":"string","enum":["CLIENT","BRAND","ADMIN"]},"name":{"type":"string","maxLength":100,"minLength":0},"pictureURL":{"type":"string","maxLength":500,"minLength":0},"linkOfficial":{"type":"string","maxLength":255,"minLength":0}},"required":["email","name","password","role"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ProductDocument":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"brandId":{"type":"integer","format":"int64"},"brandName":{"type":"string"},"brandPictureUrl":{"type":"string"},"brandIsVerified":{"type":"boolean"},"productTypeName":{"type":"string"},"category":{"type":"string"},"price":{"type":"number"},"generalFit":{"type":"string"},"gender":{"type":"string"},"color":{"type":"string"},"available":{"type":"boolean"},"rating":{"type":"number","format":"double"},"linkProduct":{"type":"string"},"imageUrls":{"type":"array","items":{"type":"string"}},"tags":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"}}},"ProductSearchResponse":{"type":"object","properties":{"searchedProducts":{"type":"array","items":{"$ref":"#/components/schemas/ProductDocument"}}}},"ProductElasticSearchResponse":{"type":"object","properties":{"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductDocument"}},"currentPage":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int32"},"totalResults":{"type":"integer","format":"int64"},"pageSize":{"type":"integer","format":"int32"}}},"BrandPublicResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"pictureURL":{"type":"string"},"linkOfficial":{"type":"string"},"followers":{"type":"integer","format":"int32"},"rating":{"type":"number","format":"double"},"isVerified":{"type":"boolean"}}}}}}